aura-backend/requirements.txt
# --- Core Server ---
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
requests==2.31.0

# --- Database (PostgreSQL & ORM) ---
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1

# --- Data Validation & Settings ---
pydantic==2.6.3
pydantic-settings==2.2.1

# --- Auth & Security ---
python-jose==3.3.0
bcrypt==4.0.1
passlib==1.7.4

# --- Image Processing & Utils ---
# QUAN TRỌNG: Phải dùng numpy < 2.0 để tương thích với TensorFlow & ONNX
numpy==1.26.4
opencv-python-headless==4.9.0.80
pillow==10.2.0
cloudinary==1.39.0

# --- Email & Reports ---
email-validator==2.1.1
fastapi-mail==1.4.1
reportlab==4.1.0

# --- AI Runtime (CPU Optimized) ---
onnxruntime-gpu==1.17.1

# --- AI Development (Training) ---
# Dùng bản CPU để nhẹ Docker (Nếu server có GPU thì đổi thành 'tensorflow')
tensorflow==2.15.0
tf2onnx==1.16.1